moving s3 testcases to new file and reverting urllib version
