+++ setup.py
@@ -10,7 +10,7 @@ with codecs.open(path.join(HERE, 'README.md'), encoding='utf-8') as f:
 
 INSTALL_DEPS = ['kubernetes==9.0.0',
                 'requests>=2.20.0',
-                'urllib3==1.22',
+                'urllib3==1.24.2',
                 'boto3==1.14.28',
                 ]
 TEST_DEPS = [ 'pytest',
